Updated project folder structure -> moved some classes to core folder -> Updated imports
Removed unused whitespaces
Removed UI class (Not used)
