+++ ait_ui/__init__.py
@@ -1,5 +1,9 @@
 
-from .__version import __version__
-from .__version import get_version
-from .ui import UI
-from .session import Session
+from .__version import __version__, get_version
+from . import app
+
+__all__ = [
+    "__version__",
+    "get_version",
+    "app"
+]
