+++ ait_ui/elements/file.py
@@ -1,9 +1,9 @@
 import os
-from .element import Element
-from ..app import server, web_socket, web_request
 import tempfile
-from ..session import Session
 import threading
+import shutil
+
+from ..core import Element
 
 class File(Element):
     def __init__(self,id = None,value = None,multiple = False, save_path = None,on_upload_done = None, autoBind=True):
@@ -14,16 +14,16 @@ class File(Element):
         self.attrs["name"] = "file"
         self.attrs["multiple"] = multiple
         self.cls("file")
-        self.style("display", "none")        
-        self.save_path = save_path        
+        self.style("display", "none")
+        self.save_path = save_path
         self.events["file-upload-started"] = self.upload_started
         self.events["change"] = self.on_change
         self.on_upload_done = on_upload_done
         self.uploaded_file_path = None
         self.uploaded_file_name = None
         self.timeout = 10
-        self.counter = 0        
-    
+        self.counter = 0
+
     def get_client_handler_str(self, event_name):
         if event_name in ["input","change"]:
             return f" on{event_name}='clientEmit(this.id,this.files[0],\"{event_name}\")'"
@@ -35,8 +35,7 @@ class File(Element):
 
     def upload_done(self):
         if self.save_path is not None:
-            save_file_path = os.path.join(self.save_path, self.uploaded_file_name)            
-            import shutil
+            save_file_path = os.path.join(self.save_path, self.uploaded_file_name)
             shutil.copyfile(self.uploaded_file_path, save_file_path)
             print("File saved to", save_file_path)
             os.remove(self.uploaded_file_path)
@@ -48,7 +47,7 @@ class File(Element):
             print("upload_started", id, file["file_name"])
             self.uploaded_file_path = os.path.join(tempfile.gettempdir(), file["uid"])
             self.uploaded_file_name = file["file_name"]
-            
+
         self.counter += 1
         if self.counter > self.timeout:
             self.counter = 0
@@ -57,7 +56,7 @@ class File(Element):
             print("upload timeout")
             #TODO: add error handling here
             return
-        
+
         if os.path.exists(self.uploaded_file_path):
             # check if the file is used by another process
             try:
