+++ ait_ui/elements/input.py
@@ -1,4 +1,5 @@
-from .element import Element
+from ..core import Element
+
 class Input(Element):
     def __init__(self, id=None, value=None, type="text", placeholder=None, step=None, autoBind=True):
         super().__init__(id=id, value=value, autoBind=autoBind)
