--- ait_ui/ui.py
@@ -1,5 +1,0 @@
-class UI:
-    def __init__(self):
-        pass
-
-
