+++ examples/basic_controls.py
@@ -7,8 +7,8 @@ sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), os.path.
 #----------------------------------------
 
 from ait_ui import app
-from ait_ui.elements import Element, Elm, Row, Col, Label, Radio, Slider, Input, Check, Text
-from ait_ui.component import Component
+from ait_ui.elements import Row, Col, Label, Radio, Slider, Input, Check, Text
+from ait_ui.core import Component, Elm
 
 class BasicControls(Component):
     def __init__(self, id=None, autoBind=True, **kwargs):
