+++ examples/component_example/comp_header.py
@@ -6,8 +6,8 @@ import os
 sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), os.path.pardir)))
 #----------------------------------------
 
-from ait_ui.elements import Element , Header , Text , Image
-from ait_ui.component import Component
+from ait_ui.elements import Header, Text, Image
+from ait_ui.core import Component, Element
 
 class Comp_Imaginer_Header(Component):
     def __init__(self, id=None, autoBind=True, **kwargs):
