+++ examples/example_comp_radio.py
@@ -6,9 +6,8 @@ import os
 sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), os.path.pardir)))
 #----------------------------------------
 
-from ait_ui.elements import Element,Elm
-from ait_ui.component import Component
 from ait_ui import app
+from ait_ui.core import Component, Elm
 from examples.component_example.comp_radio import Comp_Radio
 
 class CompRadio(Component):
