+++ examples/file.py
@@ -6,10 +6,11 @@ import os
 sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), os.path.pardir)))
 #----------------------------------------
 
-from ait_ui import app
-from ait_ui.elements import Element, Elm, File, Label, Image,Button
-from ait_ui.component import Component
 import random
+
+from ait_ui import app
+from ait_ui.elements import File, Label, Image,Button
+from ait_ui.core import Component, Elm
 
 class MyApp(Component):
     def __init__(self, id=None, autoBind=True, **kwargs):
