+++ examples/session-test.py
@@ -5,9 +5,9 @@ import sys
 import os
 sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), os.path.pardir)))
 #----------------------------------------
-from ait_ui import app, UI
-from ait_ui.elements import Element, Elm, Text, Button, Row
-from ait_ui.component import Component
+from ait_ui import app
+from ait_ui.elements import Text, Button, Row
+from ait_ui.core import Component, Elm
 
 class myApp(Component):
     def __init__(self, id=None, autoBind=True, **kwargs):
